containers: a list's length follows the node index width instead of u32
`ListHead::len` was `u32` in both crates, which put a fixed four-billion ceiling
on a single list whatever the configuration was, which is the kind of cap a
63-bit id family exists to remove. A list's length is bounded by the node arena's
population, so `N::Index` is the width the quantity lives in, and it is free at
both configurations (the header is padded to the same size at 31-bit ids).

The increments were unchecked `+=`. They cannot overflow today, because a length
is bounded by the arena and `N::MAX` is half of `N::Index::MAX` (the id gives its
top bit to the capture tag), but an unchecked `+=` would turn any future
divergence into a wrapped length, where `len()` and `iter().count()` disagree and
every caller trusting the O(1) header reads the wrong count. `grown` and the
splice path now use `checked_incr`/`checked_add` and panic.

`ListArena::len` returns `N::Index`; callers wanting a plain count widen with
`IndexLike::as_usize`. `EClasses::use_list_len` does exactly that. Widening is
sound and this is the right place for it: the value leaves the container to be
compared against another use-list length when choosing a merge survivor, and is
never stored. Narrowing would have been the bug, since `as u32` on a wider config
truncates a real count and picks the wrong survivor with no diagnostic.

`layout_parity` now asserts the header at both id widths, and the circular-list
property test runs at `DenseId63`.

Verus: 1501 verified, 0 errors.

+++ containers-conformance/tests/list_arena_differential.rs
@@ -65,8 +65,14 @@ fn list_arena_trace(seed: u64, steps: usize) {
                 let pv: Vec<u32> = p.iter(PList::new(l)).map(|e| e.raw()).collect();
                 let vv: Vec<u32> = v.iter(VList::new(l)).map(|e| e.raw()).collect();
                 assert_eq!(pv, vv, "step {step}: iter({l}) diverged");
+                // Compared without a cast on either side: both `len`s return the node
+                // arena's `N::Index`, which at a 31-bit id is `u32` on both crates. The
+                // absence of a conversion here is itself part of the parity claim — if
+                // one side's cached count went back to a fixed width while the other
+                // followed the id family, this would stop type-checking rather than
+                // silently agreeing at small lengths.
                 assert_eq!(
-                    p.len(PList::new(l)) as usize,
+                    p.len(PList::new(l)),
                     v.len(VList::new(l)),
                     "step {step}: len({l}) diverged"
                 );
